support/requirements.txt
django == 3.2
djangorestframework>= 3.10.0,<3.12
psycopg2>= 2.8.0,<2.9.3
postgres==4.0
celery==5.2
requests>=2.25,<2.26
djangorestframework-simplejwt==4.6.0
django-redis>=4.10
django-filter>=2.4.0
django-environ>=0.4.5
